ARM: dts: exynos: move ldo25 regulator node from dtsi to dts files
According to https://github.com/LineageOS/android_kernel_samsung_smdk4412
/blob/cm-14.1/arch/arm/mach-exynos/midas-power.c#L276-L713, particularly
line 472, we see that the min and max microvolt for the ldo25 regulator
on i9300 and i9305 (which have CONFIG_REGULATOR_MAX77686 enabled) should
be 3300000, not 2800000 (which it is for the Reference Device PQ aka
trats2), so ldo25 data needs to be moved out of the galaxy-s3 dtsi and
into i9300, i9305 and trats2 dts files.  Also the name for ldo25 for
RDPQ/trats2 is different.

Signed-off-by: Jack Knightly <J__A__K@hotmail.com>